pyproject.toml
[project]
name = "gcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=6.2.0",
    "matplotlib>=3.10.8",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tensorflow>=2.20.0",
    "tqdm>=4.67.1",
    "wordcloud>=1.9.5",
]
